Extract generator to separated file
